pyproject.toml
[tool.poetry]
name = "lichtblick-analytics-engineering-case-study"
version = "0.1.0"
description = ""
authors = ["emre76 <emre_arikan@yahoo.de>"]
readme = "README.md"

[tool.poetry.dependencies] 
python = "^3.10"
pandas = "^2.2.2"
click = "^8.1.7"
google-cloud-storage = "^2.16.0"
pyarrow = "^16.0.0"
dbt-bigquery = "^1.7.7"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
